Code review changes, and a fix to the testing to temporarily deal with changing roots.
